chirp-backend/python-ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0

# AI/ML Dependencies
torch==2.1.1
torchaudio==2.1.1
torchvision==2.1.1
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.8
scikit-learn==1.3.2
librosa==0.10.1
openai-whisper==20231117
transformers==4.36.0
pillow==10.1.0

# Audio processing
soundfile==0.12.1
scipy==1.11.4

# Optional: OpenSMILE for prosody features
# opensmile==2.5.0

# Utilities
requests==2.31.0
aiofiles==23.2.0
python-json-logger==2.0.7

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1